Extract decorators for govspeak and HTML guides
