+++ app/decorators/guide_decorator.rb
@@ -6,39 +6,17 @@ class GuideDecorator < Draper::Decorator
   end
 
   def title
-    @title ||= case content_type
-               when :govspeak
-                 content_document.headers.find { |header| header.level == 1 }.try(:text)
-               else
-                 content_document.css('h1').first.try(:text)
-               end
+    fail 'GuideDecorator subclasses must implement title'
   end
 
   def content
-    @content ||= case content_type
-                 when :govspeak
-                   content_document.to_sanitized_html.html_safe
-                 else
-                   guide_content.html_safe
-                 end
+    fail 'GuideDecorator subclasses must implement content'
   end
 
-  private
-
-  def content_type
-    object.content_type
-  end
-
-  def guide_content
-    object.content
-  end
-
-  def content_document
-    @content_document ||= case content_type
-                          when :govspeak
-                            Govspeak::Document.new(guide_content)
-                          else
-                            Nokogiri::HTML(guide_content)
-                          end
+  def self.for(guide)
+    case guide.content_type
+    when :govspeak then GovspeakGuideDecorator.new(guide)
+    when :html then HTMLGuideDecorator.new(guide)
+    end
   end
 end
