+++ spec/decorators/shared_examples_for_a_guide_decorator.rb
@@ -1,0 +1,27 @@
+RSpec.shared_examples 'a guide decorator' do
+  describe 'delegates' do
+    let(:id) { 'foo' }
+    let(:slug) { '/foo' }
+    let(:description) { 'The test description' }
+
+    let(:decorator) { described_class.new(instance_spy(Guide, id: id, slug: slug, description: description)) }
+
+    describe '#id' do
+      subject { decorator.id }
+
+      it { is_expected.to eq(id) }
+    end
+
+    describe '#slug' do
+      subject { decorator.slug }
+
+      it { is_expected.to eq(slug) }
+    end
+
+    describe '#description' do
+      subject { decorator.description }
+
+      it { is_expected.to eq(description) }
+    end
+  end
+end
